Update FindPath in 2D version
When we calculate a path to a neighbor, we either need to update that neighbor's cost, or add it to the queue. The old code seems to try and update the cost of the current node instead of the neighbor, which does not seem right? It also very rarely causes a path to not be found, even though it exists and the end node was visited as a neighbor but never added to the queue.